refactor(kiteregisterbackend): refactor to clean architecture
refactor to various project to replicate clean architecture
https://github.com/ardalis/CleanArchitecture
