+++ backend/Core/Entities/BaseEntity.cs
@@ -1,0 +1,7 @@
+﻿namespace KiteRegister.Core.Entities
+{
+    public abstract class BaseEntity
+    {
+        public virtual int Id { get; protected set; }
+    }
+}
