+++ backend/KiteRegisterBackend.sln
@@ -3,7 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30204.135
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KiteRegisterApi", "KiteRegisterApi\KiteRegisterApi.csproj", "{E679F815-EB39-482F-9066-EFCC2CFF8414}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApi", "WebApi\WebApi.csproj", "{E679F815-EB39-482F-9066-EFCC2CFF8414}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "Core\Core.csproj", "{F83F5DD3-0F86-435E-846C-7E971ECD1F27}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "Infrastructure\Infrastructure.csproj", "{3E14634F-CF00-4958-B902-9C10BD42E4C4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +19,14 @@ Global
 		{E679F815-EB39-482F-9066-EFCC2CFF8414}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E679F815-EB39-482F-9066-EFCC2CFF8414}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E679F815-EB39-482F-9066-EFCC2CFF8414}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F83F5DD3-0F86-435E-846C-7E971ECD1F27}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F83F5DD3-0F86-435E-846C-7E971ECD1F27}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F83F5DD3-0F86-435E-846C-7E971ECD1F27}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F83F5DD3-0F86-435E-846C-7E971ECD1F27}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3E14634F-CF00-4958-B902-9C10BD42E4C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3E14634F-CF00-4958-B902-9C10BD42E4C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3E14634F-CF00-4958-B902-9C10BD42E4C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3E14634F-CF00-4958-B902-9C10BD42E4C4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
