+++ backend/WebApi/KiteEndpoints/KiteDto.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace KiteRegisterApi.Models
+namespace KiteRegister.WebApi.KiteEndpoints
 {
     public class KiteDto
     {
