+++ backend/WebApi/Startup.cs
@@ -1,4 +1,4 @@
-using KiteRegisterApi.Models;
+using KiteRegister.Infrastructure.Data;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace KiteRegisterApi
+namespace KiteRegister.WebApi
 {
     public class Startup
     {
